Migrate toolbar logic to a new file
